requirements.txt
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
sniffio==1.3.1
upstash-redis==1.3.0
urllib3==2.3.0
Werkzeug==3.1.3
ytmusicapi==1.10.2
